# booking-web-app-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking Web App API",
        "description": "API for the Booking Web App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/merchant/register": {
            "post": {
                "summary": "Register a new merchant",
                "operationId": "df66b138071cc5867743ad378cb5c557",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "merchant@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "example": "ABC Transport"
                                    },
                                    "license_number": {
                                        "type": "string",
                                        "example": "LIC123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Merchant registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/merchant/listings": {
            "post": {
                "summary": "Create a new listing",
                "operationId": "45ddc4d8ac5d50bd40d4c65b5ff5384b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Morning Bus to City"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "bus",
                                            "flight",
                                            "train"
                                        ],
                                        "example": "bus"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 25.5
                                    },
                                    "total_seats": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "start_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-09-02T08:00:00Z"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Downtown Station"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Listing created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/merchant/bookings": {
            "get": {
                "summary": "Get merchant's bookings",
                "operationId": "4219baefb43c7d18bcea088c9a77613c",
                "responses": {
                    "200": {
                        "description": "List of bookings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/merchant/stats": {
            "get": {
                "summary": "Get merchant statistics",
                "operationId": "2bf67b64e09b5592d55ac9aeef1ce9e4",
                "responses": {
                    "200": {
                        "description": "Merchant stats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}